requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-genai==0.1.0 